server/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin 
sqlalchemy 
psycopg2-binary 
python-jose
 redis
  rq
pinecone-client
